Fix ViewLayer resync wrongly assuming success in some cases.
`BKE_view_layer_synced_ensure` would report success and clear the 'out
of sync' flag of the viewlayer, even if the call to `BKE_layer_collection_sync`
could not actually perform the resync (e.g. because resync is forbidden
by one or more calls to `BKE_layer_collection_resync_forbid`).

While logically fairly bad, this issue did not seem to have any
practical consequences. So would rather not  backport this to the 5.0 beta
branch.

This fix also reveals at least one place where the usage of
`BKE_view_layer_synced_ensure` and related is muddy and not ideal: the
Scene's `foreach_id' code. This is patched as best as possible in this
commit, but is something that will have to be properly fixed at some
point most likely.

Also add some documentation to this API - although the whole thing needs
a real reafctor at some point still, name-wise and organization-wise.