UI: Add "context menu" popover to vertical tabs
The goal of this patch is to be able to add relatively unobtrusive
popover button to the tabs region in properties editor.
To prevent adding much visual noise, the idea was to only draw little
arrow without any background or outline. The button is grayed out,
when all tabs are visible.

To only draw arrow, is achieved by setting UILayout.emboss to NONE
or NONE_OR_STATUS. This may not be most correct way of doing this,
but it looks closest to expected behavior with using current API.

The button was slightly misaligned. To align it, I have done it in very
hacky way - rcti of the button is translated by needed offset if it is
inside RGN_TYPE_NAV_BAR region type.