Fix #147653: Deleting a scene while playing it back causes a crash
The `ScreenAnimData` which is saved to the
animation timer on the screen during playback
stores a pointer to the scene that is playing.
When deleting a scene, the next time the timer
is triggered this will result in an invalid pointer
dereference.

This fixes the crash by stopping the playback when
a scene is deleted.