Fix #147642: Cycles: Random crashes of OSL tests
The root of the issue is caused by non-thread-safe access to global
variable in the liboslexec. A fuller description on when it happens
and a fix has been reported upstream:

  https://github.com/AcademySoftwareFoundation/OpenShadingLanguage/pull/2029

This change side-steps the issue on Blender side in a way that does
not require re-compilation of libraries (with a patch applied).
Once all platforms have an OSL library with the fix included the
workaround could be removed.