Fix: Timeline versioning conflict with pre-2.80 files
Move the versioning code that switches the Timeline editor to Dope Sheet
from the 'first' to the 'after linking' versioning stage.

There was already versioning code that effectively turns the Dope Sheet
back into a Timeline, for pre-2.80 files, and this runs in the 'after
linking' stage. The 5.0 versioning code now correctly runs after this.