Fix #147680: Crash when voxel size is too small
These grid creation functions can return a grid without data when
certain parameters are invalid. The "set_output" function of the
geometry node execution parameters asserts that the data is
not null, so I assume this is invalid and causes crashes later on.

---

I'm not sure that this is really the best fix, maybe SocketValueVariant
should be able to handle null grids properly. But it does fix the bug,
so maybe it's best for now.