Fix #147683: Menu Close on Leave Granularity
This PR adds more flexibility for the changes introduced in #145523.
That change allowed the new preference to turn off the "mouse move
quit" behavior all over. This PR instead uses the existing flag
USER_MENU_CLOSE_LEAVE as intended so we can allow this in some areas
despite the user setting. This means mostly just adding this flag if
USER_MENU_CLOSE_LEAVE is set, except set always for color picker and
for the small operator confirms (like object delete). Basically this
allows us to pick and choose which areas should always close on leave.