Fix #136783: Scaling sometimes causes images to shift
Scaling images in the compositor sometimes causes images to shift, this
is because the realization domain was computed using the center of the
original image before scaling. To fix this, we make sure the
transformation is actually centered.

This is really apparent and thus no tests need to be updated and this
should not be backported for safety.