Armature: Elongate tiny bones instead of deleting them
Blender cannot handle bones with (near) zero length. Prior to this
commit such bones were deleted when exiting Armature Edit mode. Now they
are kept and elongated so that they are numerically stable (or at least
they should be, given the threshold to the length that was already in
place).

To avoid the elongation from impacting the position of child bones, they
are disconnected from the tiny bone.

Apart from that it's quite nice for users that Blender no longer
silently deletes bones, this is also useful for the USD importer, as it
can import bones and expect them to exist afterwards (see  #147048).