wayland: Rework Wayland surface state management
Wayland surface state are updated using double buffered atomic commits
and each committed state is assumed to effectively describe a complete
image of how a window should be presented.

During the mapping of a window, this is done in two steps:

First create an initial "empty" configuration, where empty here means
without any pixel content or metadata that describes any pixel content
(such as regions, viewports, etc). This includes window management
requests, such as maximization, etc, which will effect how the
compositor will configure the the window that si being mapped.

Second, the compositor will emit a first toplevel configure event
(emitted via the equivalent libdecor handle configure callback), which
will contain information about how the window will be initially mapped.

A Wayland client may not commit any actual content to the surface until
having received the initial configure event, and it should always commit
content that respects the sent configuration, to avoid graphical
glitches.

Following this, future configure events will change the window
configuration, and Blender must act accordingly, and if there are
multiple configure events, Blender only need to respond to the latest
one it sees.

With libdecor, it's the same, with the only difference being that
Wayland events arrive mostly unchanged via the libdecor API, and that
libdecor will signal when it wants Blender to commit another frame, to
update any decorations.

This commit changes the configuration and commit handling to handle this
according to the expectations of Wayland compositors.

Firstly, it rearranges the mapping of a GHOST_WindowWayland to create the
initial empty commit atomically, before attempting to draw any content.

Secondly, it moves the 'initial_configure_seen' to the generic Wayland
level of the window struct, is the same semantics apply to
xdg_toplevel+decoration manager.

Lastly, during GHOST_WindowWayland construction, there is only one
single synchronous "wait", which is for both when using libdecor and
without, which is for the initial configuration, as it's not possible to
proceed with an undefined window size until having received it.