Fix #147759: Refactor handling of Scene deletion.
Deleting the current Scene ID is complex, since a lot of 'global' data
(in the WM or context e.g.) need to be carefully updated, to ensure that
they are not null and that there is still a valid active scene).

There are also rules that e.g. prevent deleting the last local Scene.

Existing code was handling these topics in several, not-fully-consistent
places (mainly `ED_scene_delete` and `rna_Main_scenes_remove`).

Library deletion from the Outliner would also delete the active scene it
it was linked from that library.

So this commit refactors the whole scene deletion process to allow
factorizing most of its logic, while still allowing for custom handling
of the ID deletion itself:
- Add `ED_scene_replace_for_deletion`, gathering and replacing most of
  the logic from both `ED_scene_delete` and `rna_Main_scenes_remove`.
  Also use this in Outliner's ID deletion code.
- Add `BKE_scene_find_replacement` to factorize the logic to choose a
  replacement scene. In addition to existing logic (use previous or next
  scene in Main), also allow to pass a 'validate' callback to check for
  other conditions for the candidate replacing scene.
- For outliner, in code checking and gathering all IDs to be deleted
  (`outliner_id_delete_tag` and related), keep track of current scene
  being deleted, and the best replacement one. Then ensure the active
  scene replacement before effectively deleting the selected IDs.

Note: This change is way too involved for 5.0 at this point, considering
that the issue has been around for many years & releases already.

Will try to do a simpler fix for 5.0 and active LTSs, likely simply
preventing Library deletion from the Outliner if the active scene is
from the deleted library.