WIP: File Browser System Root List
This PR adds a file system root listing that is a parent of all drives
so that we can browse to all volumes and shares without needing to have
the left side region open. This should be particularly useful when
using file browser embedded in the main window as with the Video
Editing workspace.