Nodes: Replace node socket labels with label_fn declaration callback
Node sockets can get "label"s which override their "name" for UI
display. These labels are not the same as node labels in that they are
not user-defined, but just a way to change the label dynamically. They
use an update function that edits the label string data in place.

This commit removes the label system and data for most nodes, and
replaces it with a callback `label_fn` in node declarations.

The main benefit is that it allows using arbitrary translation
contexts for the labels, if they clash with other messages elsewhere
in the UI.

Labels are still kept in node DNA because the new label system only
works on nodes that use declarations. The legacy combine and separate
color texture nodes do not.

Fixes #105113.

Alternative to !105690.