Fix: VSE: Crash when rendering recursive strip
When a "sequencer-type" scene strip inside of scene A points at another
VSE timeline in scene B that has a scene strip that points back to scene
A, this is classified as a "recursive strip" by
`relations_check_scene_recursion`. Notably, if other renderable strips
exist, the function will not care and the render will continue anyways.

So, to trigger a crash, all you have to do is create a recursive strip,
place any other strip in the timeline bounds (e.g. color strip), and
attempt to render the animation.

Instead of preventing the user from rendering if there are recursions
detected, this fixes the issue by moving around the recursive check that
already existed for "sequencer-type" scene strips.

Now, the check also applies to *any* scene strip, and the marked scene
is the *enclosing* scene of a scene strip rather than the scene the
strip points to (so that we can immediately notice a scene recursion).

NOTE that before, the VSE preview somehow managed to render recursive
scenes just fine without any crashes, so this PR technically makes that
behavior no longer possible. I am not sure if this is behavior we want
to keep, or if it was unexpected anyways.