wayland: Default to buffer scale of 1
When mapping a window, it's not possible to derive a buffer scale by
looking at the output's it entered, as a surface won't enter an output
until it's mapped.

While it's possible to get a preferred buffer scale before that, we're
currently not using it, due to an early out in
outputs_changed_update_scale() when there are no associated outputs.
We'll address that later, but for now, just default to 1 so that we can
reliably map a new window.