Fix: Simplify import in VR add-on
In the Viewport VR Preview, an unusual and complicated import
manipulates sys.path to import a class from a file that is already
importable anyway.

There is no documented reason for this, and the common way to subclass
an RNA type in python is by using its `bpy.types....` registered class,
_not_ by directly subclassing the python type used to declare/register
the parent type.