Fix: report progress in Fix to Camera operator
Use Blender's `ProgressReport` class to report progress of the Fix to
Camera operator. Without this, it can seem to hang Blender on heavy
scenes.

-----------

@ideasman42 I added this based on [your comment in #145414][1].

[1]: https://projects.blender.org/blender/blender/pulls/145414#issuecomment-1721809

Not sure if this is the right way to go, though. Even when I pass
`context.window_manager` to the `ProgressReport` class, it doesn't
actually show anything in the window, at least not for me on Linux.

If I do **no** pass the WM to the class, the progress report is showing
correctly on the terminal (as in, prints one line that smoothly counts
up to 100%). However, when I do pass the WM, it doesn't show this smooth
increase, and only prints `Progress:   0.00%`.

My suspicion is that the `ProgressReport` class is a bit buggy. It's
also quite a complex thing that supports various use cases (including
nested reporting), while on the other hand I couldn't find any code that
actually uses this.