EEVEE: Support different dimensions vector sockets
This patch adds support for the dimensions property of vector sockets,
support 2D and 4D sockets in addition to the currently supported 3D one.
This just involves assigning the correct GPU type when constructing the
GPU materials from nodes.

The only node that uses 2D sockets for now is the Radial Tilling node,
so we just need to update its interface.

This depends on !148033 to be actually useful, since the current
implicit conversion rules for float2 makes this pretty useless when used
with texture coordinates and such.