Compositor: Support Radial Tilling node
This patch adds support for the Radial Tilling node in the compositor.
The node already existed in the compositor, but it immediately crashed
Blender, so we dropped support for it in 5841cb21f7 until proper support
is added.

Support was added in 5517941a60, 97e3fb0e03, a2146839d7, and df8683ec2a.
So we now enable it again by reverting 5841cb21f7.