Fix #147374: Crash when adding freeze frame
Adding freeze frame to first or last key was not handled well. Maybe I
should say, that it was not handled at all.

To make this work the code needed just support for one sided clamping of
offset and different method of adding freeze keys to first / last key.
Because there has to be 2 method of adding keys, this was moved to
separate function, that returns pair of keys, so the actual logic is
more readable.