Fix #146387: VSE: Match active scene to new sequencer scene
The problem:
The issue referenced is technically not a bug, but rather user confusion
about how sequencer scenes work in VSE 5.0.

However, from a user-side, there seems to be a regression in behavior
between the Video Editing app template (where sequencer scene equals the
scene) and a new Video Editing workspace (where the new button in the
selector can cause them to differ). As a result, the VSE preview may
appear to ignore any tweaking from the properties editor.

A solution to the problem should ensure that story tools features are
available at any time to the intermediate user, while avoiding any big
disruptions for the absolute beginner. Ideally, basic editing in the VSE
should always be straightforward and never require prior
assumptions/knowledge about how the system works.

This PR is one of the potential fixes that was proposed at the most
recent VSE module meeting:

Switch the active scene to any newly created sequencer scene, ensuring
the two are equal at initialization. This has its own potential
downsides, but it ensures that adjusting sequencer render properties
always works as expected.