Fix #146387: VSE: Add Video Editing workspace to default startup.blend
NOTE that this is one potential solution to the issue laid out below,
the other is in its own PR.

The problem:
The issue referenced is technically not a bug, but rather user confusion
about how sequencer scenes work in VSE 5.0.

However, from a user-side, there seems to be a regression in behavior
between the Video Editing app template (where sequencer scene equals the
scene) and a new Video Editing workspace (where the new button in the
selector can cause them to differ). As a result, the VSE preview may
appear to ignore any tweaking from the properties editor.

A solution to the problem should ensure that story tools features are
available at any time to the intermediate user, while avoiding any big
disruptions for the absolute beginner. Ideally, basic editing in the VSE
should always be straightforward and never require prior
assumptions/knowledge about how the system works.

This PR is one of the potential fixes that was proposed at the most
recent VSE module meeting:

Add the Video Editing workspace to the default startup.blend file,
setting the sequencer scene to be the active scene.
This serves multiple purposes:
- It solves the above issue for new users. Users who wish to benefit
  from new Story Tools features can still click the righthand icon to
  create new sequencer scenes.
- It makes the VSE more prominent and discoverable for beginners.
- It gives us the chance to resave the startup.blend file with
  compression enabled, going from ~800KB -> ~120KB.

Like in #145802, I have recorded the process of changing the startup
file since it is a sensitive procedure.