Fix: Cycles random crashes in HIP-RT BVH
This change addresses random crashes reported by users on HIP-RT.

Currently, BLAS memory is allocated in the device class (in build_bvh)
but released in the BVHRT destructor. To avoid the possibility of the
GPU accessing unmapped memory, the safer approach is to release this
memory in the device class instead.

This follows the same logic as OptiX and Metal RT BVH.