Paint: Tools: Add API to get the last active brush asset of a brush type
This adds an API to get the last active brush asset of a specific
brush type. This is useful for when the user wants to activate
a specific tool (with a specific brush type) from another tool.
One such example is the Grease Pencil eraser, which can be
temporarily activated from the Draw tool.