Fix #148112: Always show colored main axis in ortho view in viewport
In Overlay Next, the grid drawing was rewritten, however the behaviour
of drawing for main axis is not the same as pre 4.3 where in ortho view
in the viewport, XYZ axis will always be colored regardless of overlay
options, which can be considered beneficial since it's much easier to
visually identify the view direction.

This fix restores original behaivour.