Fix: VSE: Don't show a menu when adding a new sequencer scene
Pressing the `New` button in the VSE would pop up a menu
showing the different options for creating a scene. This includes
the duplication options (e.g. "Full Copy" creates a full copy of the
active scene). This doesn't make sense in the context of the
sequencer if there isn't an existing sequencer scene.

This changes the invoke function of `SCENE_OT_new_sequencer_scene`
to check if a sequencer scene exists, and if it doesn't,
just run the execute function with the "New" setting.