Fix wrap_coord for out-of-range numbers
Numbers outside the ±MAXINT range could produce out-of-range wrap
results and the wrong side for extend results. This is due to these
values turning into 0x80000000 which is an unexpected negative
number, and also due to floored_fmod turning inf into nan.

New version of wrap_coord(u,size,wrap) is also at least 5% faster
by avoiding the switch statement for in-range values, and that
an if statement based on the sign of a number is much faster than
a second % operator.