UI Experiment: Window Decorations
This is not what you think it is. Wayland might need an option to have
us draw our own titlebar. This is just a hacky proof of concept for us
to look at and play with. This does not actually remove the OS chrome,
just adds our own below it, copying the window style. Only drawing at
a fixed size, assuming 1X scale and low DPI. Does most things except
window dragging as I don't think we have ghost code to do this from
here.