UI: Outliner Rename Auto-Scroll Only at Edges
In the Outliner when you rename an item in a long list it automatically
scrolls to vertical center. It can be disorienting to have the item you
are interacting with suddenly move far away. AFAIK the reason for this
is to allow continual sequential renaming of items in a list that is
longer than the available space. Otherwise it is possible that the next
item is just a sliver that is too small to select without having to
scroll it manually into view. This PR alters this behavior so that
auto-scrolling only occurs for items at the very top and bottom, always
leaving a full extra row exposed. This allows editing almost all items
in place without anything moving, with only minimal movement for the
top and bottom items.