GPU: Shader: Parse Create Info and generate placeholder macros
This is the first step towards getting rid of the
implicit shader source ordering.

Create info files are now treated as source files,
pulling their dependencies.

Note that this patch is not generating the resources macros.
This is for future work. Thus, we still pull the create info
dependencies (aka `typedef_source`).