Set initial flags based on dope sheet mode
Set these based on the initial dope sheet mode:

| Flag               | Dope Sheet | Timeline  |
|--------------------|------------|-----------|
| Channelbox visible | Yes        | No        |
| Summary            | Expanded   | Collapsed |

Note that these flags are only set when the space data is created, based
on the dope sheet mode (technically, the editor sub-type) the space data
is created for. So this is in effect when you, for example, switch the
3D View to either Timeline or Dope Sheet.

Once the space data is created, it is shared between Timeline and Dope
Sheet (because Blender doesn't create a separate space data for each
editor sub-type).