Fix #148300: Cannot interact with viewport behind region overlap
Regression from 6536fd80fc.

Apparently the "current" (`View2D.cur`) visible rectangle of the View2D
data may be bigger than the "total" (`View2d.tot`) one. E.g. when the
sidebar doesn't have enough content to fill the entire region height,
the current rectangle is the full region height, but the total one just
the height of the contents. This isn't obvious, and I thought this
wasn't the case when reviewing.

We could clamp the current rectangle by the total rectangle. But I find
going back to the total rectangle, plus checking the total rectangle as
earlier to be more readable.