Fix #148338: renaming a bone can rename channel in unassociated action
If an action was assigned to an action constraint, its channel paths
would get renamed even on completely unrelated bone renames on other
armatures.

The underlying issue was that the code that renames the channels from an
action on an action constraint didn't verify that the renamed bone and
the action constraint were from the same object.  Combined with the fact
that the constraint channel remapping code (which called that code) just
does a dumb loop over all the objects in the scene, this meant that any
bone rename on any armature would rename channels on any action assigned
to an action constraint.

This fixes the issue by only performing that channel remapping for an
action constraint's action if both the renamed bone and constraint are
from the same object.