Fix hidden nodes with many sockets improper offset during NWAlignNodes
For a while now, NWAlignNodes has had problems with hidden nodes that
have many sockets. Because of the fact that when a node is hidden,
its sockets are centered around the its "original titlebar" (IE the
titlebar it would have if all of its sockets were hidden). This change
takes the socket offsets into account, and the size of this implied
titlebar.