UI: Theme: Add Node Outline theme setting
The subtle outline introduced in 916f0afd45 is too strong in light
themes.

Add a theme setting instead so it's possible to adjust it or make
it fully transparent.

This is also an improvement in accessibility because contrast can be
enhanced in regular nodes, not just the active and selected ones.

Fixes #147810