UI: Rename "Amount" in various places to Number, Factor, Percentage
In various contexts, quantities are called "Amount", whether they are
countable or uncountable. The latter usage is sometimes considered
incorrect in English [Wiktionary], and is translated in a different
way in some languages.

Depending on context, "Amount" is better stated as a number, a factor,
or a percentage, which is more precise and avoids the translation
conflicts.

Reported by Alexandr Fatih in #43295.

[Wiktionary]: https://en.wiktionary.org/wiki/amount