UI: Add an optional `item_search_property` to `uiLayout.prop_search`.
Besides a few special cases, `prop_search` would always use the `RNA
struct name` property of the collection items, as sources for its search
strings.

This commit allows to instead specify a dedicated string property in the
items of the collection, which content will be used to generate the
searched items data.

Addresses part of the problem described in #145682.