I18n: Do not translate label in tree view
Tree views can contain user-created data. Currently the labels for
each row is translated in the tree view itself, but it shouldn't be in
the general case. For example, Light Linking objects' names were
translated.

This commit disables this translation, it is up to each tree view to
decide if it should translate its labels. To compensate for this,
translation is restored inside the node tree interface