UI: Close Blocks With Custom Event Loop on Mouse Press Outside
For blocks with custom event processing, if mouse click (or wheel
up/down) occurs while outside of the block, and when not "Close Menu
on Leave" enabled, close the block and return UI_RETURN_OK. This PR
gives the same result as #147844 but more generically within
ui_handle_menu_event. Note that the only thing I can see that assigns
a block_event_func is color picker.